[Mac] Replace obsolete POSIX functions
bcopy() and bzero() were obsoleted in POSIX 1.2001 and removed in
POSIX 1.2008.

Also set _POSIX_C_SOURCE=200809L macro so we don't accidentally
use them again.

+++ src/Core/GameController.m
@@ -514,7 +514,7 @@ static void RemovePreference(NSString *key)
 		// No AddOns at all. Show the first existing AddOns folder (paths are in order of preference, etc...).
 		BOOL		pathIsDirectory;
 		NSString	*path = nil;
-		uint		i = 1;
+		NSUInteger	i = 1;
 		
 		while (i < [[ResourceManager rootPaths] count])
 		{
